Promote draft-store to a directory module
